requirements.txt
groq
fal-client
python-dotenv